feat(236-01): decompose maps service facade
